DBZ-180 Added a new ConnectorEngine that runs one or more connectors for a single consumer
The new `ConnectorEngine` is an improvement over the older `EmbeddedEngine` in that it
supports multiple connectors as well as dynamically adding, removing, starting, and stopping
individual connectors. When a connector is added (or deployed) to the engine, a custom
classloader and callback handler can be optionally provided. The classloader is used to
load the connector classes, while the callback handler is used to report all state
transitions, such as when the connector starts, stops, fails, and when the connector's
tasks are started and stopped. The `ConnectorEngine` performs most operations asynchronously,
but in these cases the methods always return a `Future` that can be used to wait for or
determine if the asynchronous work has completed.

Like the `EmbeddedEngine`, the `ConnectorEngine` is designed to handle a single consumer
of all events generated by the deployed and running connectors. However, unlike the
`EmbeddedEngine` that expects the consumer to behave synchronously (records the offsets
of each event as soon as the consumer method processes the event), the `ConnectorEngine`
only records each events offset when the consumer explicitly calls the event's `commit()`
method. The consumer is still required to commit each event in the same order in which
it received the events, but the `ConnectorEngine` has some safety mechanisms to prevent
accidental rewinding of the offsets should the events be committed in the wrong order.