DBZ-250 Making status update interval configurable;
Using the default of PG (10 seconds) helps to reduce CPU utilization, but it causes our tests to run much longer,
as the connector will only stop after the next status update.
So we are using the default in general now, only for the tests the default is overridden with a much smaller value.