DBZ-1079 Add a sortable ID to each message
CURRENTLY WORK IN PROGRESS! Still needs:
- mongo integration test
- mysql integretion test

This commit adds a new optional ID that can be used to order the
messages.

It introduces a new configuration option, `ORDERED_ID_PROVIDER` that appends
a new field `order_id` which can be used to order transactions.

The default implementation doesn't do anything, to make this be opt-int
behavior.