DBZ-1228 MySQL connection with client authentication does not work
The SSL properties were being reset, and thus the authentication was failing on the second time.
When debezium connects to the database the first time in order to test the connections, the required parameters are all set. And thus the test passes.
After that the shutdown method is called that clears these properties, and the consecutive connection fails.

DBZ-1228: MySQL connection with client authentication does not work

Create and pass a socket factory that provides they keystore config

DBZ-1228: MySQL connection with client authentication does not work

Prevent forcing of SSL. Use SSL only if the required properties are set.

DBZ-1228: MySQL connection with client authentication does not work

Added better exception messages