DBZ-1244 Detecting unreplicated purged GTIDs;
The purpose of this change is to provide the same functionality for
when_needed snapshot for GTID enabled database like for non-GTID.
The non-GTID just relies on missing replication log to detect the
new snapshot needs to be triggered. This does not work in GTID
environment as we need to compare the GTIDs that are available on
the server with those we have already processed - stored in the
offsets.

If there is a set of GTIDs that were not replicated by Debezium
and at the same time have been already purged then we detect the
situation and trigger a new snapshot when required.

We thus

* Take all GTIDs ever processed by the server
* Filter out GTIDs that were processed by Debezium
* Compare them with purged GTID set
* If the interesction is not empty we know we have not replicated
transaction that have already been removed from the server