DBZ-1065 Use separate connection for metadata
Adaptive streaming does not work in SQL Server JDBC driver when more
than one query are executed on the same database connection.
As we are querying additional metadata while iterating the change table
the whole result set would be read into the memory if the metadata would
not be queried on a separate connection.