DBZ-1684 Make sure publication initialized prior to slot
This is to avoid the need to restart the connector when streaming begins due to
an error that the publication can't be found.  It would seem based on testing
that since we create the replication slot early on when it does not exist, then
when streaming starts its confused about trying to marry a newly created
publication to an slot.  By creating the publication before the slot, this
works around the problem.