DBZ-1452 Add the possibility to insert "op" and "ts_ms" field to the flattened record
We needed to add the operation and timestamp field from the payload struct to the record. Therefore, we adapted the ExtractNewRecordState SMT by adding two boolean configuration options (one for "op" and one for "ts_ms").  It could be a nice addition to include this permanently in the SMT.

The solution is based on pull-request #968, which added the ability of adding fields from the source struct. The addSourceFields and makeUpdatedSchema functions were broken down into smaller parts.
The original functionality is completely preserved. However, we had to add a prefix to the fields in order to prevent duplicate field names (ts_ms for example). As prefix, we chose the name of the struct the field comes from. Any suggestions on this are welcome!