DBZ-1942 Oracle connector integration tests poll JMX
The oracle integration tests use a number of Thread.sleep() calls to wait for
the snapshotting phase to finish before testing streaming changes but sometimes
this can be unreliable, especially on a slow running db. Instead of sleeping,
polling the jmx metrics to watch for when the snapshot phase ends is preferable.