DBZ-2583 MySQL connector - ignore statement-based logs
Current mysql debezium connector fails once it sees an event generated
with statement or mixed based replication.

As users and programs can call SET SESSION binlog_format='STATEMENT'; as
they wish, they are able to break debezium connector. Only MySQL 8.0.14
is fixing this, but not everyone is on that version.

To prevent connector crashing in such cases, there should be an option
to log these issues and continue. That would be similar to processing
inconsistent schema (on the setting level).