DBZ-2897 Extended RecordCommitter interface to handle updated source offsets.
Extended the DebeziumEngine interface to include a new method for marking a
record processed with updated source offsets. Implemented the new interface
method in both ConvertingEngineBuilder and EmbeddedEngine. EmbeddedEngine
creates a new SourceRecord with the updated sourceOffsets and passes this
to the original markProcessed method. Added a unit test in EmbeddedEngineTest
that verifies that the file offset storage contains the updated source offset’s
partition. Added a revapi ignore case to prevent failure from a method added
to an interface without a default method. Updated docs for Debezium Engine
that describe the new functionality.