DBZ-3268 Use dedicated connection for XStream API
Once a connection is bound to XStream API, issuing any other type of query
or operation can lead to failures due to concurrent use of the connection
by Debezium and XStream's native methods.  This commit adds a dedicated
connection used specifically by XStream to avoid this concurrency issue,
so that data types such as TIMESTAMP WITH LOCAL TIME ZONE works.