[docs] Restore downstream ModuleID for config properties topic
Previously, I removed the downstream ModuleID declaration for the Connector properties topic on L1947  because of a linking problem. I now believe that the problem resulted from errors in the way that the downstream content was structured, so I've restored the Module ID and the link to it on L1642.