[docs] fix typo in mongodb connector example
Hello,

First, thanks for making this project open source. It is such a great piece of software.
While I was playing with it and try to configure a mongodb connector, I noticed that the json example to configure a mongodb connector is not valid. There is an unnecessary comma at the end of the payload.

Here is a minor contribution that could help and could save some time for the future users.