DBZ-3328 Use object arrays to write column values
Previously we used a wrapper class, LogMinerColumnValueImpl, to store
both the column name and value; however this introduced some overhead
in object allocation.  This change mostly focused on the removal of
that implementation and its interface in favor of creating object
arrays up-front and writing the column values directly.

Included as well as the following changes:

* Deprecated LogMinerColumnValue, impl, and wrapper
* Removed SEQUENCE# from LogMiner query, no longer needed
* Adapted tests to runtime changes with object arrays