[docs] Move anchor ID, correct xref link
The inclusion of two consecutive anchor links breaks processing when files are fetched and split for downstream use. Testing downstream shows that rather than deleting the  second anchor ID `[[mongodb-tailing-the-oplog]]` that is associated with the heading *Streaming changes*, move it to follow the heading. 
Also inserted missing `debezium` string into the anchor ID in the downstream xref that links to `default-names-of-kafka-topics-that-receive-debezium-mongodb-change-event-records`