DBZ-4033 - A new hash version has been added so that it can be used to have the same hash value in different places. The byte array also has only the actual value.
Formatting was done and github user information was fixed.

DBZ-4033 PR comments have been fixed.

An addition to the documentation was made for V2.

DBZ-4033 A version has been added so that there can be backwards compatibility. The byte problem in the hash has also been fixed in version 2.

DBZ-4033 ObjectOutputStream writing extra characters.

Using ObjectOutputStream in the hash process causes data corruption.Because it performs a hash operation by adding abnormal characters to the front of the given value, the columns are corrupted and transmitted.Instead, a string conversion was performed and aObjectOutputStream writing extra characters. byte array was obtained.