DBZ-4067 Correctly calculate mining session boundaries
It is not guaranteed that the flush strategy will force all LGWR buffer entries
to be written to the redo logs in which case the mining boundaries must then be
calculated based on the lastProcessedScn rather than the current batch endScn.
This prevents event loss between mining sessions when this timing scenario
occurs.