DBZ-4366 basic support for DBMS_LOB package
Make less assumptions about the order of SELECT_LOB_LOCATOR and
LOB_WRITE events when a LOB value is constructed. The current approach
also supports interleaved LOB creation (piecemeal creation of 2 large
LOB values in different rows or columns at the same time).

There is still no support for partial LOB manipulations (trimming a LOB,
appending a couple of bytes, overwriting just a couple of bytes in the
middle, ...). To properly support that, a wholly different approach is
needed.

The internal representation of BLOB values for LogMiner has been changed
from `HEXTORAW('...')` and `BlobChunkList` to a simple byte array for
efficiency.