[docs] Remove  annotations from community topic
Remove  modularization comments and comment out xref in bulleted list in the topic Using Avro as the payload format.

Despite the topic being conditionalized for upstream use only, it's preceded by the following downstream modularization comments and the ModuleID that is defined in the comments is incorrectly included as a cross-reference  early in the file. The xref results in a build error downstream. 

// Type: concept
// Title: Using Avro as the payload format in {prodname} outbox messages
// ModuleID: using-avro-as-the-payload-format-in-debezium-outbox-messagescomments were added 

For now, due to the imminent upstream releases for 1.7x and 1.8.x, I'm making the changes in the 1.6 branch only.