DBZ-5204 Avoid incorrect connection close in PgOutputMessageDecoder
Since #3454, the same PostgreSQL connection is used for repeated
PgOutputMessageDecoder instantiations. Therefore the
PgOutputMessageDecoder#close method should not close this connection,
because it belongs to the instantiator. (Previously each instantiation
of PgOutputMessageDecoder created its own PostgreSQL connection, and so
the close method did have the responsbility of closing this connection.)

The observable symptom of this bug was a Debezium crash in the following
scenario:

  1. Create Debezium connector.

  2. Add a new type:
     CREATE TYPE enum1 AS ENUM ('val1', 'val2');

  3. Add a new table:
     CREATE TABLE enum_type (f1 enum1);

  4. Wait for Debezium to crash because the metadata connection it would
     use to fetch the metadata about `enum1` is closed.