DBZ-5181 Specify a taskId and maxTasks to run multiple EmbeddedEngines
A Debezium Server implementation may run multiple engines to parallelize
work between them. To enable multi-tasking an engine should be assigned
an ID and the provide total number of tasks. Using this information, it
can identify the slice of work from the source connector.

Debezium Server accepts a new parameter `debezium.server.threads` to
specify the number of engines to run. It also passes taskId and maxTasks
to the engine through properties.

The engine parameterizes offset storage if required by the taskId.